% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation.R
\name{cite_sources}
\alias{cite_sources}
\title{Cite data sources}
\usage{
cite_sources(x)
}
\arguments{
\item{x}{An object of class \code{tbl} produced by \code{\link[=get_data]{get_data()}}}
}
\value{
A tibble containing information on data source and relevant terms of
use.
}
\description{
Cite data sources
}
\examples{
epidata <- get_data_epidemiology(ccode = "UK", adm = 0)
cite_sources(x = epidata)

}
\author{
Ernest Guevarra
}
