% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_oxcovid19.R
\name{connect_oxcovid19}
\alias{connect_oxcovid19}
\title{Initiate a connection to the OxCOVID19 PostgreSQL database}
\usage{
connect_oxcovid19(
  dbname = "covid19",
  host = "covid19db.org",
  port = 5432,
  user = "covid19",
  password = "covid19"
)
}
\arguments{
\item{dbname}{Name of database. Default is \code{"covid19"}. Unless
OxCOVID19 Project changes database details, this should be kept as
default.}

\item{host}{Name of database host. Default is \code{"covid19db.org"}. Unless
OxCOVID19 Project changes database details, this should be kept as
default.}

\item{port}{Port specifications. Default is \code{5432}. Unless OxCOVID19
Project changes database details, this should be kept as default.}

\item{user}{User specifications. Default is \code{"covid19"}. Unless
OxCOVID19 Project changes database details, this should be kept as default.}

\item{password}{Password specifications. Default is \code{"covid19"}. Unless
OxCOVID19 Project changes database details, this should be kept as
default.}
}
\value{
A PostgreSQL connection to OxCOVID19 database
}
\description{
Initiate a connection to the OxCOVID19 PostgreSQL database
}
\examples{
connect_oxcovid19()

}
\author{
Ernest Guevarra
}
