% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_sources}
\alias{data_sources}
\title{Data sources for OxCOVID19 Database}
\format{
A named list of 6 tibbles of data sources for \strong{OxCOVID19 Database}\tabular{ll}{
   \strong{Table name} \tab \strong{Description} \cr
   \strong{epidemiology} \tab Tibble of sources for epidemiology table \cr
   \strong{government_response} \tab Tibble of sources for government response table \cr
   \strong{country_statistics} \tab Tibble of sources for country statistics table \cr
   \strong{mobility} \tab Tibble of sources for mobility table \cr
   \strong{weather} \tab Tibble of sources for weather table \cr
   \strong{administrative_division} \tab Tibble of sources for administrative division table \cr
}
}
\source{
\url{https://covid19.eng.ox.ac.uk/data_sources.html}
}
\usage{
data_sources
}
\description{
Data sources for OxCOVID19 Database
}
\examples{
data_sources

}
\keyword{datasets}
