% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_structures}
\alias{data_structures}
\title{Data structures for OxCOVID19 Database}
\format{
A named list of 6 tibbles of data structures for \strong{OxCOVID19 Database}\tabular{ll}{
   \strong{Table name} \tab \strong{Description} \cr
   \strong{epidemiology} \tab Tibble of description of fields for epidemiology table \cr
   \strong{government_response} \tab Tibble of description of fields for government response table \cr
   \strong{country_statistics} \tab Tibble of description for fields for country statistics table \cr
   \strong{mobility} \tab Tibble of description for fields for mobility table \cr
   \strong{weather} \tab Tibble of description for fields for weather table \cr
   \strong{administrative_division} \tab Tibble of description for fields for administrative division table \cr
}
}
\source{
\url{https://covid19.eng.ox.ac.uk/data_schema.html}
}
\usage{
data_structures
}
\description{
Data structures for OxCOVID19 Database
}
\examples{
data_structures

}
\keyword{datasets}
