% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Get remote table metadata}
\usage{
get_metadata(x)
}
\arguments{
\item{x}{An object of class \code{tbl_lazy} produced by \code{\link[=get_table]{get_table()}}}
}
\value{
A named list of remote table metadata
}
\description{
Get remote table metadata
}
\examples{
x <- get_table(con = connect_oxcovid19(), tbl_name = "epidemiology")
get_metadata(x)

}
\author{
Ernest Guevarra
}
