% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{list_tables}
\alias{list_tables}
\title{List out names of tables available from OxCOVID19 Database}
\usage{
list_tables(con = connect_oxcovid19())
}
\arguments{
\item{con}{A PqConnection class object specifying the PostgreSQL connection
to \strong{OxCOVID19 Database}. This is produced by \code{\link[=connect_oxcovid19]{connect_oxcovid19()}}
function}
}
\value{
A vector of table names available from \strong{OxCOVID19 Database}
}
\description{
List out names of tables available from OxCOVID19 Database
}
\examples{
list_tables()

}
\author{
Ernest Guevarra
}
