% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_sources}
\alias{data_sources}
\title{Data sources for OxCOVID19 Database}
\format{
A named list of 5 tibbles of data sources.
\describe{
  \item{\code{epidemiology}}{Tibble of sources for epidemiology table}
  \item{\code{government_response}}{Tibble of sources for government
    response table}
  \item{\code{country_statistics}}{Tibble of sources for country statistics
    table}
  \item{\code{mobility}}{Tibble of sources for mobility table}
  \item{\code{weather}}{Tibble of sources for weather table}
}
}
\source{
\url{https://covid19.eng.ox.ac.uk/data_sources.html}
}
\usage{
data_sources
}
\description{
Data sources for OxCOVID19 Database
}
\examples{
data_sources

}
\keyword{datasets}
