% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_structures}
\alias{data_structures}
\title{Data structures for OxCOVID19 Database}
\format{
A named list of 6 tibbles of data structures.
\describe{
  \item{\code{epidemiology}}{Tibble of description of fields for epidemiology
    table}
  \item{\code{government_response}}{Tibble of description of fields for
    government response table}
  \item{\code{country_statistics}}{Tibble of description for fields for
    country statistics table}
  \item{\code{mobility}}{Tibble of description for fields for mobility table}
  \item{\code{weather}}{Tibble of description for fields for weather table}
  \item{\code{administrative_division}}{Tibble of description for fields for
    administrative division table}
}
}
\source{
\url{https://covid19.eng.ox.ac.uk/data_schema.html}
}
\usage{
data_structures
}
\description{
Data structures for OxCOVID19 Database
}
\examples{
data_structures

}
\keyword{datasets}
