% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{list_fields}
\alias{list_fields}
\title{List out field names of specific tables available from OxCOVID19 Database}
\usage{
list_fields(con = connect_oxcovid19(), tab = list_tables())
}
\arguments{
\item{con}{A PqConnection class object specifying the PostgreSQL connection
to OxCOVID19 Database. This is produced by \code{connect_covid19}
function}

\item{tab}{A character vector specifying table name/s to extract field names
from. Must be a value included when a call to \code{list_names} is issued.
Default is character vector of all table names specified by
\code{list_names}.}
}
\value{
A named list of field names per specified table available from
  OxCOVID19 Database
}
\description{
List out field names of specific tables available from OxCOVID19 Database
}
\examples{
list_fields()

}
\author{
Ernest Guevarra
}
