% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxcovid19.R
\docType{package}
\name{oxcovid19}
\alias{oxcovid19}
\alias{oxcovid19-package}
\title{An R API to the Oxford COVID-19 Database}
\description{
The OxCOVID19 Project \url{https://covid19.eng.ox.ac.uk} aims to increase
our understanding of the COVID-19 pandemic and elaborate possible strategies
to reduce the impact on the society through the combined power of statistical,
mathematical modelling, and machine learning techniques. The OxCOVID19
Database is a large, single-centre, multimodal relational database consisting
of information (using acknowledged sources) related to COVID-19 pandemic.
This package provides an R-specific interface to the OxCOVID19 Database based
on widely-used data handling and manipulation approaches in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/como-ph/oxcovid19}
  \item \url{https://como-ph.github.io/oxcovid19}
  \item Report bugs at \url{https://github.com/como-ph/oxcovid19/issues}
}

}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID})

Authors:
\itemize{
  \item Chris Mercado \email{cegmercado@gmail.com} (\href{https://orcid.org/0000-0002-5588-9543}{ORCID})
  \item Robert Medina \email{jcmedina1@up.edu.ph} (\href{https://orcid.org/0000-0001-6647-5453}{ORCID})
  \item Alexander Zarebski
}

}
\keyword{internal}
